change page to decision as default page.
