+++ components/Header/NavLinks.tsx
@@ -31,7 +31,7 @@ const NavLinks = ({ links, listStyle }: NavLinksProps) => {
                                 className={
                                     'buttonText ' +
                                     (link.href.includes(router.asPath)
-                                        ? 'text-primary font-bold underline underline-offset-8'
+                                        ? 'text-primary font-bold underline underline-offset-4'
                                         : 'text-neutral-700 dark:text-neutralDark-150')
                                 }
                             >
